(Oanda) Computes decimal quantities

+++ cpp/sandbox/src/oanda/tao_bot/max_affordable_quantity.cpp
@@ -5,7 +5,7 @@
 #include "convert_price.cpp" // convert_price
 #include "tao_bot.h"         // Oanda::TaoBot
 
-int Oanda::TaoBot::max_affordable_quantity() {
+double Oanda::TaoBot::max_affordable_quantity() {
   const double dollars_per_unit = convert_price(1.0, base_currency(), "USD");
 
   return this->account_balance.margin_buying_power / dollars_per_unit;
